pythonic-raqa-lc-app/requirements.txt
chainlit==0.7.700
langchain==0.0.350
openai==0.27.8
tiktoken==0.4.0
chainlit==0.7.700
faiss-cpu==1.7.4
pinecone-client
arxiv
pypdf